Connect functions to html

+++ main/search/forms.py
@@ -3,6 +3,6 @@ from django import forms
 class LocationForm(forms.Form):
     location1 = forms.CharField(label='location1')
     location2 = forms.CharField(label='location2')
-    radius = forms.FloatField(label='radius')
-    ratio = forms.FloatField(label='ratio')
-    budget = forms.FloatField(label='budget')
+    radius = forms.IntegerField(label='radius')
+    ratio = forms.IntegerField(label='ratio')
+    budget = forms.IntegerField(label='budget')
